+++ R/lagrange.R
@@ -55,7 +55,7 @@
 #' lagrange(mod, list(parnum[1:5], parnum[c(1, 5)]))
 #'
 #' }
-lagrange <- function(mod, parnum, SE.type = 'Oakes', type = 'central', ...){
+lagrange <- function(mod, parnum, SE.type = 'Oakes', type = 'Richardson', ...){
     fn <- function(par, mod, pn, dat, model, parprior, PrepList, large, sv, ObJeCtIvE, MG, group, ...){
         sv2 <- sv
         sv2$value[pn] <- par
@@ -74,7 +74,8 @@ lagrange <- function(mod, parnum, SE.type = 'Oakes', type = 'central', ...){
         ret <- extract.mirt(mod2, 'logLik')
         ret
     }
-    grads <- function(mod, pn, dat, model, parprior, PrepList, large, sv, ObJeCtIvE, type = 'central', MG, group, ...){
+    grads <- function(mod, pn, dat, model, parprior, PrepList, large, sv, ObJeCtIvE,
+                      type = 'Richardson', MG, group, ...){
         par <- sv$value[pn]
         g <- numerical_deriv(fn, par, mod=mod, pn=pn, dat=dat, model=model, parprior=parprior,
                              PrepList=PrepList, large=large, sv=sv, MG=MG, group=group,